Deduplicate query client execution and split API from implementation.
Centralize RawExecuteQueryRequest construction and client retry loops,
and split client_query into builders, internal dispatch, stream facade, and exec.
